fix: disable swagger in prd
