+++ .github/workflows/prd-api-deploy.yml
@@ -37,7 +37,7 @@ jobs:
         id: login-ecr
         uses: aws-actions/amazon-ecr-login@v1
       - name: Build and Push Docker image to Amazon ECR
-        run: ./gradlew :api:jib -Djib.from.image=openjdk:17-jdk-slim -Djib.to.image=${{secrets.ECR_IMAGE}} -Djib.to.auth.username=AWS -Djib.to.auth.password=${{ steps.ecr-login.outputs.ecr-password }} -Djib.to.tags=prd -Djib.container.jvmFlags="-Dspring.profiles.active=prd,-Dfile.encoding=UTF-8" -Djib.container.environment="DATASOURCE_URL=${{ secrets.PRD_DATASOURCE_URL}},DATASOURCE_USERNAME=${{ secrets.PRD_DATASOURCE_USERNAME }},DATASOURCE_PASSWORD=${{ secrets.PRD_DATASOURCE_PASSWORD }},GOOGLE_OAUTH_CLIENT_ID=${{ secrets.GOOGLE_OAUTH_CLIENT_ID }},GOOGLE_OAUTH_CLIENT_SECRET=${{ secrets.GOOGLE_OAUTH_CLIENT_SECRET }},GOOGLE_OAUTH_REDIRECT_URI=${{ secrets.GOOGLE_OAUTH_REDIRECT_URI}}"
+        run: ./gradlew :api:jib -Djib.from.image=openjdk:17-jdk-slim -Djib.to.image=${{secrets.ECR_IMAGE}} -Djib.to.auth.username=AWS -Djib.to.auth.password=${{ steps.ecr-login.outputs.ecr-password }} -Djib.to.tags=prd -Djib.container.jvmFlags="-Dspring.profiles.active=prd,-Dfile.encoding=UTF-8" -Djib.container.environment="AES_128_KEY=${{ secrets.AES_128_KEY }},ACCESS_TOKEN_EXPIRE=${{ secrets.ACCESS_TOKEN_EXPIRE }},REFRESH_TOKEN_EXPIRE=${{ secrets.REFRESH_TOKEN_EXPIRE }},ACCESS_TOKEN_SECRET=${{ secrets.ACCESS_TOKEN_SECRET }},REFRESH_TOKEN_SECRET=${{ secrets.REFRESH_TOKEN_SECRET }},DATASOURCE_URL=${{ secrets.PRD_DATASOURCE_URL}},DATASOURCE_USERNAME=${{ secrets.PRD_DATASOURCE_USERNAME }},DATASOURCE_PASSWORD=${{ secrets.PRD_DATASOURCE_PASSWORD }},GOOGLE_OAUTH_CLIENT_ID=${{ secrets.GOOGLE_OAUTH_CLIENT_ID }},GOOGLE_OAUTH_CLIENT_SECRET=${{ secrets.GOOGLE_OAUTH_CLIENT_SECRET }},GOOGLE_OAUTH_REDIRECT_URI=${{ secrets.GOOGLE_OAUTH_REDIRECT_URI}}"
       - name: Trigger ECS
         run: |
           aws ecs update-service --cluster ${{ secrets.AWS_ECS_CLUSTER }} --service ${{ secrets.AWS_ECS_SERVICE }} --force-new-deployment
