--- src/api/src/main/kotlin/com/tenmm/tilserver/common/config/rest/WebConfig.kt
@@ -1,15 +1,0 @@
-package com.tenmm.tilserver.common.config.rest
-
-import org.springframework.context.annotation.Configuration
-import org.springframework.web.reactive.config.CorsRegistry
-import org.springframework.web.reactive.config.WebFluxConfigurer
-
-@Configuration
-class WebConfig : WebFluxConfigurer {
-    override fun addCorsMappings(registry: CorsRegistry) {
-        registry.addMapping("/**")
-            .allowedOrigins("*")
-            .allowedHeaders("*")
-            .allowedMethods("*")
-    }
-}
