+++ src/api/src/main/kotlin/com/tenmm/tilserver/common/config/rest/WebFluxConfig.kt
@@ -3,6 +3,7 @@ package com.tenmm.tilserver.common.config.rest
 import com.tenmm.tilserver.common.security.annotation.OptionalUserArgumentResolver
 import com.tenmm.tilserver.common.security.annotation.RequiredUserArgumentResolver
 import org.springframework.context.annotation.Configuration
+import org.springframework.web.reactive.config.CorsRegistry
 import org.springframework.web.reactive.config.WebFluxConfigurer
 import org.springframework.web.reactive.result.method.annotation.ArgumentResolverConfigurer
 
@@ -13,5 +14,12 @@ class WebFluxConfig(
 ) : WebFluxConfigurer {
     override fun configureArgumentResolvers(configurer: ArgumentResolverConfigurer) {
         configurer.addCustomResolver(requiredUserArgumentResolver, optionalUserArgumentResolver)
+    }
+
+    override fun addCorsMappings(registry: CorsRegistry) {
+        registry.addMapping("/**")
+            .allowedOrigins("*")
+            .allowedHeaders("*")
+            .allowedMethods("*")
     }
 }
